Move _expand_curi and fix curi expansion of literal datatypes
